refactor: BottomNavigation을 HomeLayout에서 떼어 BottomNavigationLayout으로 승격
