+++ client/src/App.tsx
@@ -4,14 +4,17 @@ import Landing from './page/Landing';
 import RootLayout from './layout/RootLayout';
 import Home from './page/Home';
 import HomeLayout from './layout/HomeLayout';
+import BottomNavigationLayout from './layout/BottomNavigationLayout';
 
 function App() {
     return (
         <Routes>
             <Route element={<RootLayout />}>
                 <Route path="/landing" element={<Landing />} />
-                <Route element={<HomeLayout />}>
-                    <Route path="/home" element={<Home />} />
+                <Route element={<BottomNavigationLayout />}>
+                    <Route element={<HomeLayout />}>
+                        <Route path="/home" element={<Home />} />
+                    </Route>
                 </Route>
             </Route>
         </Routes>
